+++ bindings/python/ParametersHandler/src/YarpModule.cpp
@@ -1,14 +1,14 @@
 /**
- * @file Module.cpp
- * @authors Giulio Romualdi, Diego Ferigo
+ * @file YarpModule.cpp
+ * @authors Giulio Romualdi
  * @copyright 2021 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
  */
 
 #include <pybind11/pybind11.h>
 
-#include <BipedalLocomotion/bindings/ParametersHandler/Module.h>
-#include <BipedalLocomotion/bindings/ParametersHandler/ParametersHandler.h>
+#include <BipedalLocomotion/bindings/ParametersHandler/YarpModule.h>
+#include <BipedalLocomotion/bindings/ParametersHandler/YarpParametersHandler.h>
 
 namespace BipedalLocomotion
 {
@@ -16,12 +16,9 @@ namespace bindings
 {
 namespace ParametersHandler
 {
-void CreateModule(pybind11::module& module)
+void CreateYarpModule(pybind11::module& module)
 {
-    module.doc() = "Parameter handler module";
-
-    CreateIParameterHandler(module);
-    CreateStdParameterHandler(module);
+    CreateYarpParameterHandler(module);
 }
 } // namespace ParametersHandler
 } // namespace bindings